refactor: splitting code into subpackages
