+++ internal/models/comment.go
@@ -1,0 +1,5 @@
+package models
+
+type Comment struct {
+	Name, Comment string
+}
