+++ internal/models/conf.go
@@ -1,0 +1,5 @@
+package models
+
+type Conf struct {
+	AllowComments bool `yaml:"allow_comments"`
+}
